feat: markers component
