+++ frontend/apps/web/app/_components/index.ts
@@ -1,1 +1,2 @@
 export { default as Aside } from './Aside';
+export { default as Markers } from './Markers';
